refactor: ProjectMyStatusesでステータス部分をmapで回すようにし、ProjectMyStatusコンポーネントを作成した #256
