+++ frontend/src/pages/projectDetail/ProjectDetail.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState, useMemo } from 'react'
+import React, { FC, useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { resetServerContext } from 'react-beautiful-dnd'
 import styled, { css } from 'styled-components'
